refactor(backend): separate initialization of servers from app.ts
